저장시스템 기초 완료

+++ Assets/WorkSpace/ZB/Scripts/CSV/Parser.cs
@@ -7,24 +7,53 @@ namespace ZB.CSV
 {
     public class Parser
     {
-        public static List<Dictionary<string, string>> ReadSingle(string path, out string fileName)
+        public enum AccessStyle
+        {
+            Resources,
+            PersistentDataPath
+        }
+
+        public static List<Dictionary<string, string>> ReadSingle(AccessStyle accessStyle, string path, out string fileName)
         {
             List<Dictionary<string, string>> result = new List<Dictionary<string, string>>();
 
-            TextAsset sourceFile = Resources.Load<TextAsset>(path);
-            StringReader sr = new StringReader(sourceFile.text);
-            fileName = sourceFile.name;
+            string csvData = "";
+            fileName = "";
 
-            if (sr != null)
+            switch (accessStyle)
+            {
+                case AccessStyle.Resources:
+                    TextAsset sourceFile = Resources.Load<TextAsset>(path);
+
+                    if (sourceFile == null)
+                    {
+                        Debug.LogError($"Parser.ReadSingle / ReadByResources / Failed / Path : {path}");
+                        return null;
+                    }
+
+                    StringReader sr = new StringReader(sourceFile.text);
+                    fileName = sourceFile.name;
+                    csvData = sr.ReadToEnd();
+                    Debug.LogError($"Parser.ReadSingle / ReadByResources / Success");
+                    break;
+
+                case AccessStyle.PersistentDataPath:
+                    string fullPath = Path.Combine(Application.persistentDataPath, path);
+                    fileName = Path.GetFileName(fullPath);
+                    csvData = File.ReadAllText(fullPath);
+                    Debug.LogError($"Parser.ReadSingle / ReadByPersistentDataPath / FullPath : {fullPath}");
+                    break;
+            }
+
+            if (csvData != "")
             {
                 string temp = "";
                 int index = 0;
-                string text = sr.ReadToEnd();
 
                 //열 분리시킬 텍스트 떼내기
-                while (text[index + 1] != '\n')
+                while (csvData[index + 1] != '\n')
                 {
-                    temp += text[index++];
+                    temp += csvData[index++];
                 }
                 string[] textKeys = temp.Split(",");
                 index += 2;
@@ -39,12 +68,12 @@ namespace ZB.CSV
                         temp = "";
 
                         //줄바꿈 있는 텍스트일 경우
-                        if (text[index] == '\"')
+                        if (csvData[index] == '\"')
                         {
                             index++;
-                            while (text[index] != '\"')
+                            while (csvData[index] != '\"')
                             {
-                                temp += text[index++];
+                                temp += csvData[index++];
                             }
 
                             if (i + 1 >= textKeys.Length)
@@ -56,15 +85,15 @@ namespace ZB.CSV
                         //줄바꿈 없는 텍스트일 경우
                         else
                         {
-                            while (text[index] != ',' && text[index] != '\n')
+                            while (csvData[index] != ',' && csvData[index] != '\n')
                             {
-                                temp += text[index++];
+                                temp += csvData[index++];
                             }
                             index++;
                         }
                         result[result.Count - 1].Add(textKeys[i], temp);
 
-                        if (index + 1 >= text.Length)
+                        if (index + 1 >= csvData.Length)
                         {
                             endLoop = true;
                             break;
@@ -77,7 +106,7 @@ namespace ZB.CSV
             Debug.LogError("Path에 해당하는 파일없음 : " + path);
             return null;
         }
-        public static List<List<Dictionary<string, string>>> ReadMultiple(string path, out string[] fileNames)
+        public static List<List<Dictionary<string, string>>> ReadMultipleByResources(string path, out string[] fileNames)
         {
             List<List<Dictionary<string, string>>> result = new List<List<Dictionary<string, string>>>();
             List<Dictionary<string, string>> tempList;
@@ -157,7 +186,7 @@ namespace ZB.CSV
             return result;
         }
 
-        public static void WriteSingle(string path, List<Dictionary<string,string>> data)
+        public static void WriteSingle(AccessStyle accesStyle, string path, List<Dictionary<string,string>> data, bool canOverwrite = false)
         {
             string csvData = "";
 
@@ -167,7 +196,7 @@ namespace ZB.CSV
                 csvData += entry.Key + ",";
             }
                 csvData = csvData.Remove(csvData.Length - 1);
-            csvData += "\n";
+            csvData += "\r\n"; // CRLF로 변경
 
             //Values
             foreach (Dictionary<string, string> row in data)
@@ -177,37 +206,26 @@ namespace ZB.CSV
                     csvData += entry.Value + ",";
                 }
                 csvData = csvData.Remove(csvData.Length - 1);
-                csvData += "\n";
+                csvData += "\r\n"; // CRLF로 변경
             }
-
-            Debug.LogError("---------------------");
-            //아스키코드로 읽기
-            for (int i = 0; i < csvData.Length; i++)
-            {
-                Debug.LogError(i + " / " + (int)csvData[i] + " / " + csvData[i]);
-            }
-            Debug.LogError("---------------------");
 
             //경로입력
             string fullPath="";
 
-            bool isAndroid=false;
+            switch(accesStyle)
+            {
+                case AccessStyle.PersistentDataPath:
+                fullPath = Path.Combine(Application.persistentDataPath, path);
+                    break;
+            }
 
-#if UNITY_ANDROID
-            fullPath = Path.Combine(Application.streamingAssetsPath, "Resources", path, "Save.csv");
-            isAndroid = true;
-#endif
+            Debug.LogError($"Parser.WriteSingle / FullPath : {fullPath}");
 
-#if UNITY_EDITOR
-            fullPath = Path.Combine(Application.dataPath, "Resources", path + "/Save.csv");
-            isAndroid = false;
-#endif
-
-            Debug.LogError(isAndroid);
-            Debug.LogError(fullPath);
-
-            //덮어쓰기
-            File.WriteAllText(fullPath, csvData);
-        }
+            if (!File.Exists(fullPath) ||                   //새로생성조건
+                (File.Exists(fullPath) && canOverwrite))    //덮어쓰기조건
+            {
+                File.WriteAllText(fullPath, csvData);
+            }
+        }        
     }
 }
